requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.1.1
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
env==0.1.0
environs==11.2.1
exceptiongroup==1.2.2
executing==2.1.0
inflection==0.5.1
ipython==8.18.1
jedi==0.19.2
marshmallow==3.23.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
wcwidth==0.2.13